Refactor: Accomodation 엔티티 위치, 패키지 구조 수정
