+++ src/main/java/ybe/mini/travelserver/domain/room/Room.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.Comment;
-import ybe.mini.travelserver.domain.accommodation.Accommodation;
+import ybe.mini.travelserver.domain.accommodation.entity.Accommodation;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
